requirements.txt
# SPR - Sistema Preditivo Royal
# Python Backend Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Redis Cache
redis==5.0.1
hiredis==2.2.3

# Data Processing
pandas==2.1.3
numpy==1.25.2
scipy==1.11.4

# Machine Learning
scikit-learn==1.3.2
joblib==1.3.2

# HTTP Requests
httpx==0.25.2
requests==2.31.0

# Environment & Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Authentication & Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# Data Validation
email-validator==2.1.0

# Logging
loguru==0.7.2

# Date/Time
python-dateutil==2.8.2
pytz==2023.3

# Agricultural Data APIs
beautifulsoup4==4.12.2
lxml==4.9.3

# File Processing
openpyxl==3.1.2
xlsxwriter==3.1.9

# Async Support
asyncio-mqtt==0.13.0
asyncpg==0.29.0

# WhatsApp Integration Support
Pillow==10.1.0

# Financial Calculations
python-decimal==1.3

# Testing (dev)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0